requirements.txt
algorithms==0.1.4
gym==0.26.2
keras==2.7.0
matplotlib==3.5.2
more_itertools==4.2.0
numpy==1.17.4
Pillow==9.3.0
scikit_learn==1.1.3
scipy==1.7.2
stable_baselines3==1.3.0
torch==1.10.0+cu113
torchvision==0.11.1+cu113
